Fix buggy ledge grabs
found this bug in mario builder. the ledge grab function checks all collided walls. if you collide with two walls at the same time that can be ledge grabbed, even though the second wall can be "ignored" as the first wall is the one to ledge grab on, the ledge grab position will still be updated and will fuck up the grab